clean up the server side and put every thing in /server/routes => So I have to fix where we fetch. Also deleted some unnecessary dependencies (or packages)
