+++ package.json
@@ -6,28 +6,18 @@
   "type": "module",
   "scripts": {
     "start": "nodemon server/index.js",
+    "client": "cd journey-ai && npm start",
+    "server": "cd server && npm start",
+    "dev": "concurrently \"npm run server\" \"npm run client\"",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^6.6.0",
-    "@fortawesome/fontawesome-svg-core": "^6.6.0",
-    "@fortawesome/free-brands-svg-icons": "^6.6.0",
-    "@fortawesome/free-regular-svg-icons": "^6.6.0",
-    "@fortawesome/free-solid-svg-icons": "^6.6.0",
-    "@fortawesome/react-fontawesome": "^0.2.2",
-    "@google/generative-ai": "^0.14.1",
-    "@react-google-maps/api": "^2.19.3",
-    "@vis.gl/react-google-maps": "^1.1.0",
-    "axios": "^1.7.2",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "nodemon": "^3.1.4",
-    "react-redux": "^9.1.2",
-    "react-router-dom": "^6.24.1",
-    "unsplash-js": "^7.0.19",
-    "use-places-autocomplete": "^4.0.1"
+    "nodemon": "^3.1.4"
+  },
+  "devDependencies": {
+    "concurrently": "^8.2.2"
   }
 }
